Update: add more documentation to Attributes + refactor Application class into separate interface and implementation
