+++ src/Core/Dal/Attributes/Column.php
@@ -3,6 +3,9 @@ namespace App\Core\Dal\Attributes;
 
 use Attribute;
 
+/**
+ * Specify the column name to map the property to. Only usable in class properties.
+ */
 #[\Attribute(Attribute::TARGET_PROPERTY)]
 class Column
 {
