+++ src/Core/Dal/Attributes/RefType.php
@@ -3,6 +3,9 @@ namespace App\Core\Dal\Attributes;
 
 use Attribute;
 
+/**
+ * Specify the nested class to transform. Only usable in class properties.
+ */
 #[\Attribute(Attribute::TARGET_PROPERTY)]
 class RefType
 {
